Added Korumori, which as far as i know is the working version of this app.
Korumori's big difference is that it mainly uses xaml for form rendering ( rather than c# ) . This makes it a bit cleaner and sticks to mvvm design more.

The ui is reworked, i dropped the master detail page for tabbed page for now ( we only have 1 working page )

Most of this is uncommented, but i will get to fully documenting everything with xml tags next week.